Reorganized legacy UKF state estimator for 2D and 7D state space
